Put request-handling logic in `QuickRequestHandler`
